pyproject.toml
[tool.poetry]
name = "hugging-face"
version = "0.1.0"
description = "Pratice for hugging face exercises"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "hugging_face"}]

[tool.poetry.dependencies]
python = "^3.11"
transformers = {extras = ["torch"], version = "^4.34.0"}
jupyterlab = "^4.0.7"
pillow = "^10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
